# encrypt_app/static/api-spec.json
{
  "definitions": {
    "ServiceResponse": {
      "properties": {
        "data": {
          "type": "string"
        }
      },
      "type": "object"
    }
  },
  "info": {
    "title": "ASE Project",
    "version": "v1"
  },
  "paths": {
    "/create": {
      "post": {
        "description": "/create endpoint:\n\nThis API is intended to be used for creation of a password for an application. The user passes an application name and a password to this endpoint which is then saved in encrypted format.\n\nUsage: curl -X POST \"0.0.0.0:5001/create?application=app1&password=qwerty@098\"",
        "parameters": [
          {
            "description": "Application Name",
            "in": "query",
            "name": "application",
            "required": true,
            "type": "string"
          },
          {
            "description": "Plaintext Password",
            "in": "query",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    },
    "/delete": {
      "delete": {
        "description": "/delete endpoint:\n\nThis API is intended to be used for deletion of the password for an application. The user passes an application name and this endpoint deletes its corrsponding password if it exists.\n\nUsage: curl -X DELETE \"0.0.0.0:5001/delete?application=app1\"",
        "parameters": [
          {
            "description": "Application Name",
            "in": "query",
            "name": "application",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    },
    "/generate": {
      "get": {
        "description": "/generate endpoint:\n\nThis API is intended to be used for generation of a strong password for an application. The user invokes the endpoint and receives a secure password.\n\nUsage: curl -X GET \"0.0.0.0:5001/generate\"",
        "parameters": [
          {
            "description": "Password Length",
            "in": "query",
            "name": "len",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Boolean to include numbers",
            "in": "query",
            "name": "num",
            "required": false,
            "type": "boolean"
          },
          {
            "description": "Boolean to include special characters",
            "in": "query",
            "name": "spchar",
            "required": false,
            "type": "boolean"
          },
          {
            "description": "Boolean to include uppercase letter",
            "in": "query",
            "name": "caps",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    },
    "/retrieve": {
      "get": {
        "description": "/retrieve endpoint:\n\nThis API is intended to be used for retrieval of a password for an application. The user passes an application name and this endpoint decrypts the saved password and returns it in plaintext.\n\nUsage: curl -X GET \"0.0.0.0:5001/retrieve?application=app1\"",
        "parameters": [
          {
            "description": "Application Name",
            "in": "query",
            "name": "application",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    },
    "/strength": {
      "get": {
        "description": "/strength endpoint:\n\nThis API is intended to verify the strength of password taken as input from the user. The user passes a password and this endpoint provides the strength of password in terms of 0-4 score, strength label and a guessed time for cracking the given password.\n\nUsage: curl -X GET \"0.0.0.0:5001/strength?password=pa$$c0de\"",
        "parameters": [
          {
            "description": "Plaintext Password",
            "in": "query",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    },
    "/update": {
      "post": {
        "description": "/update endpoint:\n\nThis API is intended to be used for updation of a password for an application. The user passes an application name and a new password to this endpoint which is then saved in encrypted format.\n\nUsage: curl -X POST \"0.0.0.0:5001/update?application=app1&password=new_qwerty@098\"",
        "parameters": [
          {
            "description": "Application Name",
            "in": "query",
            "name": "application",
            "required": true,
            "type": "string"
          },
          {
            "description": "Plaintext Password",
            "in": "query",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ServiceResponse"
            }
          }
        },
        "tags": [
          "password manager"
        ]
      }
    }
  },

  "swagger": "2.0.0"
}